First working version
Ghost has an added AnimatorUpdateMode attached to keep track of the animator.updateMode.
Lerping for layers.
Fixed a problem with layers not being added even when they changed.
GhostManager can be used to load in a multitude of ghosts and make them run around.
Ghost recorder has some things it handles better, always performs a save at the very end now and also adds compression support.
TimedGhost is used to deserialize from json in a format that can be easily parsed for a Ghost based on time.